split get-bindings-proxy fixture tests in multiple files
